Crate substrate_test_client

Client testing utilities.

Re-exports

pub use sc_client_db;

pub use sp_consensus;

pub use sc_executor;

pub use self::client_ext::ClientExt;

pub use self::client_ext::ClientBlockImportExt;

Modules

client
Substrate Client and associated logic.

client_ext
Client extension for tests.

Structs

Backend
Disk backend.

ExecutionExtensions
A producer of execution extensions for offchain calls.

ExecutionStrategies
Execution strategies settings.

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

RpcHandlers
RPC handlers that can perform RPC queries.

RpcSession
An RPC session. Used to perform in-memory RPC queries (ie. RPC queries that don't go through the HTTP or WebSockets server).

RpcTransactionError
An error for when the RPC call fails.

RpcTransactionOutput
The output of an RPC transaction.

Storage
Struct containing data needed for a storage.

StorageChild
Child trie storage data.

TestClientBuilder
A builder for creating a test client instance.

Enums

AccountKeyring
Set of test accounts.

Ed25519Keyring
Set of test accounts.

ExecutionStrategy
Strategy for executing a call into the runtime.

Sr25519Keyring
Set of test accounts.

WasmExecutionMethod
Specification of different methods of executing the runtime Wasm code.

Traits

BlockchainEventsExt
An extension trait for BlockchainEvents.

GenesisInit
A genesis storage initialization trait.

RpcHandlersExt
An extension trait for RpcHandlers.

SyncCryptoStore
Sync version of the CryptoStore

Type Definitions

BadBlocks
Known bad block hashes.

ForkBlocks
Expected hashes of blocks at given heights.

LightBackend
Test client light database backend.

SyncCryptoStorePtr
A pointer to a keystore.