Struct substrate_test_client::TestClientBuilder

pub struct TestClientBuilder<Executor, Backend, G: GenesisInit> { /* fields omitted */ }

A builder for creating a test client instance.

Methods

impl<Block, Executor, G: GenesisInit> TestClientBuilder<Executor, Backend<Block>, G> where
Block: BlockT<Hash = <Blake2Hasher as Hasher>::Out>,

pub fn with_default_backend() -> Self

Create new TestClientBuilder with default backend.

pub fn backend(&self) -> Arc<Backend<Block>>

Give access to the underlying backend of these clients

impl<Executor, Backend, G: GenesisInit> TestClientBuilder<Executor, Backend, G>

pub fn with_backend(backend: Arc<Backend>) -> Self

Create a new instance of the test client builder.

pub fn set_keystore(self, keystore: BareCryptoStorePtr) -> Self

Set the keystore that should be used by the externalities.

pub fn genesis_init_mut(&mut self) -> &mut G

Alter the genesis storage parameters.

pub fn add_child_storage(
self,
key: impl AsRef<[u8]>,
child_key: impl AsRef<[u8]>,
value: impl AsRef<[u8]>
) -> Self

Extend child storage

pub fn set_execution_strategy(
self,
execution_strategy: ExecutionStrategy
) -> Self

Set the execution strategy that should be used by all contexts.

pub fn build_with_executor<Block, RuntimeApi>(
self,
executor: Executor
) -> (Client<Backend, Executor, Block, RuntimeApi>, LongestChain<Backend, Block>) where
Executor: CallExecutor<Block, Blake2Hasher>,
Backend: Backend<Block, Blake2Hasher>,
Block: BlockT<Hash = <Blake2Hasher as Hasher>::Out>,

Build the test client with the given native executor.

impl<E, Backend, G: GenesisInit> TestClientBuilder<LocalCallExecutor<Backend, NativeExecutor<E>>, Backend, G>

pub fn build_with_native_executor<Block, RuntimeApi, I>(
self,
executor: I
) -> (Client<Backend, LocalCallExecutor<Backend, NativeExecutor<E>>, Block, RuntimeApi>, LongestChain<Backend, Block>) where
I: Into<Option<NativeExecutor<E>>>,
E: NativeExecutionDispatch,
Backend: Backend<Block, Blake2Hasher>,
Block: BlockT<Hash = <Blake2Hasher as Hasher>::Out>,

Build the test client with the given native executor.