Crate substrate_test_client

Client testing utilities.

Re-exports

pub use self::client_ext::ClientBlockImportExt;

pub use self::client_ext::ClientExt;

pub use sc_client_db;

pub use sc_executor;

pub use sp_consensus;

Modules

client

Substrate Client and associated logic.

client_ext

Client extension for tests.

Structs

Backend

Disk backend.

ExecutionExtensions

A producer of execution extensions for offchain calls.

RpcHandlers

A running RPC service that can perform in-memory RPC queries.

RpcTransactionError

An error for when the RPC call fails.

RpcTransactionOutput

The output of an RPC transaction.

Storage

Struct containing data needed for a storage.

StorageChild

Child trie storage data.

TestClientBuilder

A builder for creating a test client instance.

WasmExecutor

An abstraction over Wasm code executor. Supports selecting execution backend and manages runtime cache.

Enums

BlockOrigin

Block data origin.

BlocksPruning

Block pruning settings.

Ed25519Keyring

Set of test accounts.

Sr25519Keyring

Set of test accounts.

WasmExecutionMethod

Specification of different methods of executing the runtime Wasm code.

Traits

BlockchainEventsExt

An extension trait for BlockchainEvents.

GenesisInit

A genesis storage initialization trait.

Keystore

Something that generates, stores and provides access to secret keys.

RpcHandlersExt

An extension trait for RpcHandlers.

Type Aliases

BadBlocks

Known bad block hashes.

ForkBlocks

Expected hashes of blocks at given heights.

KeystorePtr

A shared pointer to a keystore implementation.